// backend/lambda-instance/src/utils/enums.ts
export enum AssetTicker {
  BTC = 'BTC',
  ETH = 'ETH',
  USDC = 'USDC',
  HONEY = 'HONEY'
}

export enum LogLevel {
  INFO = 'INFO',
  WARN = 'WARN',
  ERROR = 'ERROR',
  CRITICAL = 'CRITICAL',
}

export enum SlackTag {
  MOBY = '<!subteam^S058TFTF21Y>',
  PRICE_VOLATILITY_CHECKER = '<!subteam^S08D33MGECR>',
  DEV = '<!subteam^S03Q2K1M4AY>',
}